requirements.txt
websockets>=12.0,<13.0
pymongo>=4.6.2,<5
python-dotenv>=1.0.1,<2
requests>=2.31.0,<3
watchdog>=4.0.0,<5
